refactor: align folder structure with README, group hooks, and standardize naming conventions
