--- Components/Array Example/ArrayExample.jsx
@@ -1,16 +1,0 @@
-function ArrayExample()
-{
-  const fruits=["apple","banana","pineapple"];
-
-  return(
-    <>
-      <ul>
-        {fruits.map((fruits,index)=>{
-          return <li key={index}>{fruits}</li>
-        })}
-      </ul>
-    </>
-  );
-
-}
-export default ArrayExample
