--- Components/Css Example/CssStyleExample.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export default function CssStyleExample() {
-  return (
-    <>
-       
-        <p className='test'>Charmy Vora</p>
-        <h1 id="id1">Awt lecture</h1>
-    </>
-  )
-}
