+++ Components/ImageTagExample/ImageTag.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function ImageTag() {
+  return (
+    <>
+        <h2> Image</h2>
+        <img src="./public/butterfly.png" width="100" height="100" />
+        <p>Charmy Vora</p>
+        <h1>Awt lecture</h1>
+    </>
+  )
+}
