+++ Components/MultipleCompTag/MultipleTagOne.jsx
@@ -1,0 +1,10 @@
+function MultipleTagOne()
+{
+    return(
+        <>
+            <h1>Charmy </h1>
+            <p> Vora</p>
+        </>
+    );
+}
+export default MultipleTagOne
