--- Components/MutipleComp-tag/MultipleTagOne.jsx
@@ -1,10 +1,0 @@
-function MultipleTagOne()
-{
-    return(
-        <>
-            <h1>Charmy </h1>
-            <p> Vora</p>
-        </>
-    );
-}
-export default MultipleTagOne
