--- Components/Object Example/ObjectExampleOne.jsx
@@ -1,12 +1,0 @@
-function ObjectExampleOne()
-{
-    const person={name:"Mahi",age:40};
- 
-    return(
-        <>
-            <p>First Name : {person.name}</p>
-            <p>Age : {person.age}</p>
-        </>
-    );
-}
-export default ObjectExampleOne
