--- Components/SingleComp-tag/SingleComponent.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function SingleComponent() {
-  return (
-    <div>
-        <p>Hello</p>
-    </div>
-  )
-}
