+++ Components/SingleCompTag/SingleComponent.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SingleComponent() {
+  return (
+    <div>
+        <p>Hello</p>
+    </div>
+  )
+}
