--- Components/function Example/PrintName.jsx
@@ -1,9 +1,0 @@
-function PrintName()
-{
-    const name="Raj";
-    const printName=(
-        <h5> hello {name}</h5>
-    );
-    return printName;
-}
-export default PrintName
